move a couple things around
